+++ static/app/views/insights/common/queries/types.ts
