+++ static/app/views/insights/common/utils/renameDiscoverSeries.tsx
@@ -1,4 +1,4 @@
-import type {DiscoverSeries} from 'sentry/views/insights/common/queries/useDiscoverSeries';
+import type {DiscoverSeries} from 'sentry/views/insights/common/queries/types';
 
 export function renameDiscoverSeries(
   series: DiscoverSeries,
