+++ static/app/views/insights/http/components/charts/responseCodeCountChart.tsx
@@ -10,7 +10,7 @@ import {BaseChartActionDropdown} from 'sentry/views/insights/common/components/c
 // TODO(release-drawer): Only used in httpSamplesPanel, should be easy to move data fetching in here
 // eslint-disable-next-line no-restricted-imports
 import {InsightsLineChartWidget} from 'sentry/views/insights/common/components/insightsLineChartWidget';
-import type {DiscoverSeries} from 'sentry/views/insights/common/queries/useDiscoverSeries';
+import type {DiscoverSeries} from 'sentry/views/insights/common/queries/types';
 import {getAlertsUrl} from 'sentry/views/insights/common/utils/getAlertsUrl';
 import type {AddToSpanDashboardOptions} from 'sentry/views/insights/common/utils/useAddToSpanDashboard';
 import {useAlertsProject} from 'sentry/views/insights/common/utils/useAlertsProject';
